undo outputmanager move
